requirements.txt
streamlit
langchain
langchain-groq
sentence-transformers
faiss-cpu
PyPDF2
requests
